+++ integrations/astra/src/haystack_integrations/document_stores/astra/astra_client.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import json
+from collections.abc import Generator
 from typing import Any
 from warnings import warn
 
@@ -33,7 +34,8 @@ class Response:
 class QueryResponse:
     matches: list[Response]
 
-    def get(self, key):
+    def get(self, key: str) -> Any:  # noqa: ANN401
+        """Return the value for the given key."""
         return self.__dict__[key]
 
 
@@ -50,9 +52,10 @@ class AstraClient:
         embedding_dimension: int,
         similarity_function: str,
         namespace: str | None = None,
-    ):
+    ) -> None:
         """
         The connection to Astra DB is established and managed through the JSON API.
+
         The required credentials (api endpoint and application token) can be generated
         through the UI by clicking and the connect tab, and then selecting JSON API and
         Generate Configuration.
@@ -177,17 +180,24 @@ class AstraClient:
 
         # include_metadata means return all columns in the table (including text that got embedded)
         # include_values means return the vector of the embedding for the searched items
-        formatted_response = self._format_query_response(responses, include_metadata, include_values)
+        formatted_response = self._format_query_response(
+            responses, include_metadata=include_metadata, include_values=include_values
+        )
 
         return formatted_response
 
-    def _query_without_vector(self, top_k, filters=None):
+    def _query_without_vector(self, top_k: int | None, filters: dict[str, Any] | None = None) -> list[dict[str, Any]]:
         query = {"filter": filters, "limit": top_k}
 
         return self.find_documents(query)
 
     @staticmethod
-    def _format_query_response(responses, include_metadata, include_values):
+    def _format_query_response(
+        responses: list[dict[str, Any]] | None,
+        *,
+        include_metadata: bool | None,
+        include_values: bool | None,
+    ) -> QueryResponse:
         final_res = []
 
         if responses is None:
@@ -207,7 +217,9 @@ class AstraClient:
 
         return QueryResponse(final_res)
 
-    def _query(self, vector, top_k, filters=None):
+    def _query(
+        self, vector: list[float], top_k: int | None, filters: dict[str, Any] | None = None
+    ) -> list[dict[str, Any]]:
         query = {"sort": {"$vector": vector}, "limit": top_k, "includeSimilarity": True}
 
         if filters is not None:
@@ -244,7 +256,7 @@ class AstraClient:
 
         return find_results
 
-    def find_one_document(self, find_query):
+    def find_one_document(self, find_query: dict[str, Any]) -> dict[str, Any] | None:
         """
         Find one document in the Astra index.
 
@@ -271,7 +283,7 @@ class AstraClient:
         """
         document_batch = []
 
-        def batch_generator(chunks, batch_size):
+        def batch_generator(chunks: list[str], batch_size: int) -> Generator[list[str], None, None]:
             for i in range(0, len(chunks), batch_size):
                 i_end = min(len(chunks), i + batch_size)
                 batch = chunks[i:i_end]
@@ -330,7 +342,8 @@ class AstraClient:
         ids: list[str] | None = None,
         filters: dict[str, str | float | int | bool | list | dict] | None = None,
     ) -> int:
-        """Delete documents from the Astra index.
+        """
+        Delete documents from the Astra index.
 
         :param ids: the ids of the documents to delete
         :param filters: additional filters to apply when deleting documents
@@ -352,6 +365,7 @@ class AstraClient:
     def delete_all_documents(self) -> int:
         """
         Delete all documents from the Astra index.
+
         :returns: the number of documents deleted
         """
         delete_result = self._astra_db_collection.delete_many(filter={})